% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{redgreen}
\alias{redgreen}
\title{A very simple funcion that takes in two
double arrays. One with the value to
render and another with the difference.}
\usage{
redgreen(value, diff)
}
\description{
A very simple funcion that takes in two
double arrays. One with the value to
render and another with the difference.
}
